Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.1
gunicorn>=21.2.0
SQLAlchemy>=2.0.20
pydantic>=2.5.0
email-validator>=1.3.0
pyodbc>=4.0.34
boto3>=1.35.0
python-jose[cryptography]>=3.3.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
langchain-groq>=0.0.4
pytest>=8.0
pytest-cov>=4.0
httpx>=0.24

